Move fragment downloading and uploading into a new file
Makes the code quite a bit cleaner.
